refactor: extract shared logic into loadKeplerDataset helper
