+++ package.json
@@ -20,6 +20,7 @@
     "@radix-ui/react-separator": "^1.1.2",
     "@radix-ui/react-slot": "^1.1.2",
     "@radix-ui/react-tooltip": "^1.1.8",
+    "@reduxjs/toolkit": "^2.8.2",
     "@tailwindcss/vite": "^4.0.17",
     "@tanstack/react-router": "^1.115.2",
     "class-variance-authority": "^0.7.1",
